Successfully made a single table for all reservations for both dashboard and search results. Also some visual tweaks to the status of reservatiosn and tables.
